Add draft mechanism for adding blurbs to sections
The goal is to replace the blank page at the beginning of each
section with notes on how to parse the rhythm and favorite ways
of making a baseline out of a chord (progression).

This commit wires up the mechanism to convert LaTeX peppered with
lilypond snippets into such a page.